+++ commands.py
@@ -1,3 +1,5 @@
-from arctasks import lint  # noqa
+from runcommands.commands import show_config
+
+from arctasks.base import lint  # noqa
 from arctasks.python import show_upgraded_packages  # noqa
 from arctasks.release import *  # noqa